feat: Improved navigation rail and route structure
