+++ ui/src/commonMain/kotlin/com/srfmolina/krocy/ui/presentation/feature/stock/StockScreen.kt
@@ -15,11 +15,11 @@ import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.srfmolina.krocy.domain.model.common.ConsumptionType
 import com.srfmolina.krocy.ui.presentation.common.model.ConsumptionDateUi
-import com.srfmolina.krocy.ui.presentation.common.model.TopBarConfigurationUi
-import com.srfmolina.krocy.ui.presentation.common.model.TopBarType
 import com.srfmolina.krocy.ui.presentation.feature.stock.StockViewModel.Event
 import com.srfmolina.krocy.ui.presentation.feature.stock.component.StockItemComp
 import com.srfmolina.krocy.ui.presentation.feature.stock.model.StockItemUi
+import com.srfmolina.krocy.ui.presentation.navigation.component.topbar.model.TopBarConfigurationUi
+import com.srfmolina.krocy.ui.presentation.navigation.component.topbar.model.TopBarTypeUi
 import com.srfmolina.krocy.ui.presentation.theme.KrocyTheme
 import com.srfmolina.krocy.ui.presentation.theme.spacing
 import org.koin.compose.viewmodel.koinViewModel
@@ -35,7 +35,7 @@ internal fun StockScreen(
     LaunchedEffect(Unit) {
         onChangeTopBar(TopBarConfigurationUi(
             title = "Resumen del inventario",
-            type = TopBarType.SMALL
+            type = TopBarTypeUi.SMALL
         ))
         viewModel.launchEvent(Event.Init)
     }
