+++ ui/src/commonMain/kotlin/com/srfmolina/krocy/ui/presentation/navigation/AppNavGraph.kt
@@ -3,11 +3,11 @@ package com.srfmolina.krocy.ui.presentation.navigation
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.navigation
-import com.srfmolina.krocy.ui.presentation.common.model.TopBarConfigurationUi
 import com.srfmolina.krocy.ui.presentation.feature.login.navigation.loginScreen
 import com.srfmolina.krocy.ui.presentation.feature.splash.navigation.splashScreen
 import com.srfmolina.krocy.ui.presentation.feature.stock.navigation.stockScreen
 import com.srfmolina.krocy.ui.presentation.feature.welcome.navigation.welcomeScreen
+import com.srfmolina.krocy.ui.presentation.navigation.component.topbar.model.TopBarConfigurationUi
 import kotlinx.serialization.Serializable
 
 internal fun NavGraphBuilder.appNavGraph(
