+++ ui/src/commonMain/kotlin/com/srfmolina/krocy/ui/presentation/navigation/NavigationItemUi.kt
@@ -2,7 +2,7 @@ package com.srfmolina.krocy.ui.presentation.navigation
 
 import androidx.compose.ui.graphics.vector.ImageVector
 
-internal data class KrocyNavigationItem(
+internal data class NavigationItemUi(
     val icon: ImageVector,
     val label: String,
     val contentDescription: String = label,
