+++ ui/src/commonMain/kotlin/com/srfmolina/krocy/ui/presentation/navigation/component/topbar/model/TopBarTypeUi.kt
@@ -1,0 +1,5 @@
+package com.srfmolina.krocy.ui.presentation.navigation.component.topbar.model
+
+internal enum class TopBarTypeUi {
+    SMALL, MEDIUM
+}
