+++ ui/src/commonMain/kotlin/com/srfmolina/krocy/ui/presentation/navigation/route.kt
@@ -1,0 +1,17 @@
+package com.srfmolina.krocy.ui.presentation.navigation
+
+import kotlinx.serialization.Serializable
+
+internal sealed interface KrocyRoute
+
+@Serializable
+data object SplashRoute: KrocyRoute
+
+@Serializable
+data object WelcomeRoute: KrocyRoute
+
+@Serializable
+data object LoginRoute: KrocyRoute
+
+@Serializable
+data object StockRoute: KrocyRoute
